[오류수정] 제목변경오류 수정

+++ Capstone/app/src/main/java/com/example/Capstone/TextReadActivity.java
@@ -67,7 +67,11 @@ public class TextReadActivity extends AppCompatActivity {
             file.write(data);
             file.flush();
             file.close();
-            d.delete();
+            if(!d.getName().equals(f.getName()))
+            {
+                d.delete();
+            }
+
         } catch (Exception e) {
             e.printStackTrace();
             Log.d("testTag","save fail");
